Move local chat filter from chat context to app context
